Remove default for database flag (#708)
* Change database.uri flag to default to empty

This could theoretically break deployments of the configs service that relied
on the default being set.

In practice, we anticipate very few configs deployments expecting a database
at `configs-db.weave.local`

* Only serve DB-backed stuff if DB set

In #649, we only served DB-backed endpoints for ruler if the configs API flag
was _not_ set. This turns out to be hostile to a smooth migration, as we want
to have points where we are serving _both_ sets of endpoints so we can update
the service-ui to go from pointing at one to pointing at the other.