Allow setting rule format version per user
The binary-wide rule format version flag is now removed in favor of
a per-user format version stored in the user's config JSON blob. That
JSON blob is created by the web UI and sent to the config API, so it's
now up to components outside of the Cortex backend (e.g. the Weaveworks
service UI) to set what rule format version they would like to set. If
the new JSON field for the version is omitted, it will default to 1.x
format upon unmarshaling.

This change can be rolled back safely to an older version of Cortex as
long as no 2.x rule files have been created.