Improve ClaimTokens() code for normalised case
If the ingester we are claiming from is normalising, we need to blank
out its tokens.

Simplify the rest of the claim logic since we hold a denormalised ring
all the time - the get() functions depend on it.

Stop checking for an empty denormalised ring after migration in unit
test.

Signed-off-by: Bryan Boreham <bryan@weave.works>